Fix the annoying 420 error
This bug came from a lack of understanding of the timeout. The timeout
(which is in ms and not in s as previously assumed) simply sets the
polling rate of the driver when asking for completion. The interval
started at 10ms which was so small, it was showing errors on the
screen. There is no downside to setting a high timeout since the PNA
will answer as soon as it is done averaging.